configura endpoints e services

+++ src/main/java/pt/com/hospital/domain/Service/ConsultaService.java
@@ -1,8 +1,11 @@
 package pt.com.hospital.domain.Service;
 
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PutMapping;
+import pt.com.hospital.application.DTO.mapper.ConsultaMapper;
+import pt.com.hospital.application.DTO.request.ConsultaRequest;
+import pt.com.hospital.application.DTO.response.ConsultaResponse;
 import pt.com.hospital.domain.entity.Consulta;
 import pt.com.hospital.domain.entity.Medico;
 import pt.com.hospital.domain.entity.Paciente;
@@ -10,9 +13,8 @@ import pt.com.hospital.domain.repository.ConsultaRepository;
 import pt.com.hospital.domain.repository.MedicoRepository;
 import pt.com.hospital.domain.repository.PacienteRepository;
 
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 public class ConsultaService {
@@ -27,41 +29,49 @@ public class ConsultaService {
     private PacienteRepository pacienteRepository;
 
     // agendar consulta
-    public Consulta agendaConsulta(Consulta consulta) {
-        Paciente paciente = pacienteRepository.findPacienteById(consulta.getPaciente().getId())
-                .orElseThrow(() -> new NoSuchElementException("Paciente não encontrado"));
-        Medico medico = medicoRepository.findMedicoById(consulta.getMedico().getId())
-                .orElseThrow(() -> new NoSuchElementException("Médico não encontrado"));
-
-        consulta.setPaciente(paciente);
-        consulta.setMedico(medico);
-        consulta.setStatus("Agendada");
-        consulta.setDataCadastro(LocalDateTime.now());
-        return consultaRepository.save(consulta);
+    public ConsultaResponse create(ConsultaRequest request){
+        var consulta = ConsultaMapper.toConsulta(request);
+        var novaConsulta = consultaRepository.save(consulta);
+        return ConsultaMapper.toConsultaResponse(novaConsulta);
     }
 
-    public List<Consulta> findAll(){
-        return (List<Consulta>) consultaRepository.findAll();
+    public List<ConsultaResponse> findAll(){
+        List<Consulta> consulta = (List<Consulta>)  consultaRepository.findAll();
+        List<ConsultaResponse> findConsulta = consulta.stream()
+                .map(ConsultaMapper::toConsultaResponse).toList();
+        return findConsulta;
     }
 
 
-    @PutMapping("/consulta/{id}")
-    public Consulta updateConsulta(long id, Consulta consulta) {
+    public ConsultaResponse updateConsulta(long id, ConsultaRequest consulta) {
         Consulta consultaExistente = consultaRepository.findConsultaById(id)
-                .orElseThrow(() -> new NoSuchElementException("Consulta não encontrada com id: " + id));
+                .orElseThrow(() -> new EntityNotFoundException("Consulta não encontrada com id: " + id));
 
-        consultaExistente.setMedico(consulta.getMedico());
-        consultaExistente.setPaciente(consulta.getPaciente());
+        Medico updateMedico = medicoRepository.findMedicoById(consulta.getMedicoId())
+                .orElseThrow(() -> new EntityNotFoundException("Médico não encontrado com id: " + consulta.getMedicoId()));
+        consultaExistente.setMedico(updateMedico);
+
+        Paciente updatePaciente = pacienteRepository.findPacienteById(consulta.getPacienteId())
+                .orElseThrow(() -> new EntityNotFoundException("Paciente não encontrado com id: " + consulta.getPacienteId()));
+        consultaExistente.setPaciente(updatePaciente);
+
         consultaExistente.setDataHora(consulta.getDataHora());
         consultaExistente.setMotivoConsulta(consulta.getMotivoConsulta());
         consultaExistente.setStatus(consulta.getStatus());
 
-        return consultaRepository.save(consultaExistente);
+        Consulta consultaAtualizada = consultaRepository.save(consultaExistente);
+
+        return ConsultaMapper.toConsultaResponse(consultaAtualizada);
     }
+
 
     // cancela consulta
-    public void cancelaConsulta(long id){
-        consultaRepository.deleteById(id);
+    public boolean cancelaConsulta(long id){
+        Optional<Consulta> consulta = consultaRepository.findConsultaById(id);
+        if (consulta.isPresent()){
+            consultaRepository.deleteById(id);
+            return  true;
+        }
+        return false;
     }
-
 }
